change mono repository pattern
