+++ .github/workflows/lint.yml
@@ -1,4 +1,4 @@
-on: [pull_request, push]
+on: [push]
 
 name: Lint
