+++ Cargo.toml
@@ -1,24 +1,6 @@
-[package]
-name = "halberd"
-version = "1.5.1"
-edition = "2018"
-authors = [
-	"o108minmin <o108minmin@gmail.com>",
+[workspace]
+
+members = [
+    "core",
+    "cli",
 ]
-description = "Generate Text To Speech Software Subtitle"
-repository = "https://github.com/o108minmin/halberd"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-encoding_rs = "0.8.31"
-hound = "3.4.0"
-clap = "3.1.9"
-log = "0.4.16"
-env_logger = "0.9.0"
-xml-rs = "0.8"
-time = {version = "0.3.13", features = ["macros", "formatting", "local-offset"]}
-
-[dev-dependencies]
-assert_cmd = "2.0.4"
-predicates = "2.1.1"
-tempfile = "3"
