+++ cli/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "cli"
+version = "1.5.1"
+edition = "2021"
+authors = [
+	"o108minmin <o108minmin@gmail.com>",
+]
+description = "Generate Text To Speech Software Subtitle"
+repository = "https://github.com/o108minmin/halberd"
+
+[dependencies]
+clap = "3.1.9"
+log = "0.4.16"
+env_logger = "0.9.0"
+core = {path = "../core"}
+
+[dev-dependencies]
+assert_cmd = "2.0.4"
+predicates = "2.1.1"
+tempfile = "3"
