+++ cli/output.srt
@@ -1,0 +1,8 @@
+1
+00:00:00,000 --> 00:00:03,499
+私はだみーです。よろしくお願いします。
+
+2
+00:00:03,500 --> 00:00:04,475
+こんにちわ
+
