+++ cli/tests/data/xml/normal_coefont_lazy_xml.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<fcpxml version="1.8">
+  <resources>
+    <format id="r1" name="FFVideoFormat1080p2398" frameDuration="1001/24000s" width="1920" height="1080" colorSpace="1-1-1 (Rec. 709)" />
+    <effect id="r2" name="text" />
+  </resources>
+  <library>
+    <event name="coefont_lazy">
+      <project name="coefont_lazy">
+        <sequence format="r1">
+          <spine>
+            <text-style-def id="ts1">
+              <text-style font="Helvetica" fontSize="72" fontFace="Regular" fontColor="1 0.999974 0.999991 1" alignment="center" />
+            </text-style-def>
+            <title ref="r2" offset="0/1000s" name="1" start="0/1000s" duration="1000/1000s">
+              <text>
+                <text-style ref="ts1">私はだみーです。よろしくお願いします。</text-style>
+              </text>
+            </title>
+            <title ref="r2" offset="1000/1000s" name="2" start="1000/1000s" duration="975/1000s">
+              <text>
+                <text-style ref="ts1">こんにちわ</text-style>
+              </text>
+            </title>
+          </spine>
+        </sequence>
+      </project>
+    </event>
+  </library>
+</fcpxml>
