+++ cli/tests/softalk_test.rs
@@ -8,7 +8,7 @@ use predicates::prelude::*;
 fn normal_softalk() -> Result<(), Box<dyn std::error::Error>> {
     let expected = "1\n00:00:00,000 --> 00:00:01,837\nゆっくりしていってね\n\n2\n00:00:01,838 --> 00:00:03,080\nダミーだぜ\n\n";
 
-    let mut cmd = Command::cargo_bin("halberd").unwrap();
+    let mut cmd = Command::cargo_bin("cli").unwrap();
     let assert = cmd.arg("softalk").arg("tests/data/tts/softalk").assert();
     assert.success().stdout(predicate::str::contains(expected));
     Ok(())
