+++ core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "core"
+version = "1.5.1"
+edition = "2021"
+authors = [
+	"o108minmin <o108minmin@gmail.com>",
+]
+description = "Generate Text To Speech Software Subtitle"
+repository = "https://github.com/o108minmin/halberd"
+
+[dependencies]
+encoding_rs = "0.8.31"
+hound = "3.4.0"
+log = "0.4.16"
+xml-rs = "0.8"
+time = {version = "0.3.13", features = ["macros", "formatting", "local-offset"]}
+
+[dev-dependencies]
+tempfile = "3"
