+++ core/src/xml.rs
